Soda 1.1.4
Soda 1.1.4

+++ test/src/oracle/json/tests/soda/test_OracleSodaSecurity.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2014, 2018, Oracle and/or its affiliates. 
+/* Copyright (c) 2014, 2019, Oracle and/or its affiliates. 
 All rights reserved.*/
 
 /*
@@ -26,45 +26,59 @@ import oracle.soda.rdbms.impl.SODAUtils;
 public class test_OracleSodaSecurity extends SodaTestCase {
   
   public void testSQLInjection() throws Exception {
-    
     OracleDocument metaDoc = client.createMetadataBuilder()
         .keyColumnAssignmentMethod("CLIENT")
         .mediaTypeColumnName("MediaType")
         .build();
-    OracleCollection col = dbAdmin.createCollection("testSQLInjection", metaDoc);
-    
+    OracleCollection col;
+    if (isJDCSMode())
+    {
+      col = db.admin().createCollection("testSQLInjection", null);
+    } else
+    {
+      col = db.admin().createCollection("testSQLInjection", metaDoc);
+    }
+
     String version1 = null;
     OracleDocument doc = null;
     for (int i = 1; i <= 10; i++) {
-      doc = col.insertAndGet(db.createDocumentFromString("id-" + i, "{ \"d\" : " + i + " }"));
-      
+      if (isJDCSMode()) 
+      {
+        doc = col.insertAndGet(db.createDocumentFromString("{ \"d\" : " + i + " }"));
+      } else
+      {
+        doc = col.insertAndGet(db.createDocumentFromString("id-" + i, "{ \"d\" : " + i + " }"));
+      }
+
       if(i==1)
         version1 = doc.getVersion();
     }
     
     // Tests about query operations
-    assertEquals(0, col.find().key("xxx or 1=1 ").count());
-    assertEquals(0, col.find().key("xxx ) or ( 1=1 ").count());
-    
-    assertNull(col.findOne("id\" or 1=1 or 1=\""));
-    assertNull(col.findOne("id\" ) or ( 1=1 or 1=\""));
-    
-    assertEquals(0, col.find().key("yyy' or 1=1 or 1='zzz").count());
-    assertEquals(0, col.find().key("yyy' ) or ( 1=1 or 1='zzz").count());
-    
-    assertNull(col.find().key("yyy' or 1=1 or 1='zzz").version(version1).getOne());
-    assertNull(col.find().key("yyy' ) or ( 1=1 or 1='zzz").version(version1).getOne());
-    
-    // Tests about remove operations
-    assertEquals(0, col.find().key("xxx or 1=1 ").remove());
-    assertEquals(0, col.find().key("xxx') or (1=1) --").remove());
-    
-    // Tests about replace operations
-    doc = db.createDocumentFromString("id-xxx", "{ \"d\" : \"new value\" }");
-    
-    assertEquals(false, col.find().key("xxx or 1=1 ").replaceOne(doc));
-    assertEquals(false, col.find().key("xxx') or (1=1) --").replaceOne(doc));
-    
+    if (!isJDCSMode())
+    {
+      assertEquals(0, col.find().key("xxx or 1=1 ").count());
+      assertEquals(0, col.find().key("xxx ) or ( 1=1 ").count());
+      
+      assertNull(col.findOne("id\" or 1=1 or 1=\""));
+      assertNull(col.findOne("id\" ) or ( 1=1 or 1=\""));
+      
+      assertEquals(0, col.find().key("yyy' or 1=1 or 1='zzz").count());
+      assertEquals(0, col.find().key("yyy' ) or ( 1=1 or 1='zzz").count());
+      
+      assertNull(col.find().key("yyy' or 1=1 or 1='zzz").version(version1).getOne());
+      assertNull(col.find().key("yyy' ) or ( 1=1 or 1='zzz").version(version1).getOne());
+      
+      // Tests about remove operations
+      assertEquals(0, col.find().key("xxx or 1=1 ").remove());
+      assertEquals(0, col.find().key("xxx') or (1=1) --").remove());
+      
+      // Tests about replace operations
+      doc = db.createDocumentFromString("id-xxx", "{ \"d\" : \"new value\" }");
+      
+      assertEquals(false, col.find().key("xxx or 1=1 ").replaceOne(doc));
+      assertEquals(false, col.find().key("xxx') or (1=1) --").replaceOne(doc));
+    }
     col.find().remove();
     assertEquals(0, col.find().count());
         
@@ -72,37 +86,93 @@ public class test_OracleSodaSecurity extends SodaTestCase {
       "declare pragma autonomous_transaction; "
     + "begin execute immediate 'create table tbl_by_SQLInjection ( c1 number )' end;";
     
-    col.insert(db.createDocumentFromString("id-1", createTblStr, "text/plain"));
-    assertEquals(createTblStr, new String(col.findOne("id-1").getContentAsByteArray(), "UTF-8"));
+    String[] key = new String[10];
+    if (isJDCSMode()) 
+    {
+      try 
+      {
+        doc = col.insertAndGet(db.createDocumentFromString(createTblStr));
+        fail("No Exception when doc's content is not json.");
+      }
+      catch (Exception e) 
+      {
+        String msg = e.getMessage();
+        if (msg == null)
+          fail(e.toString());
+        assertTrue(msg.contains("Error occurred during conversion of the input"));
+      }
+    } 
+
+    if (isJDCSMode()) 
+    {
+      doc = col.insertAndGet(db.createDocumentFromString("{\"originalField\" :\"" + createTblStr + "\"}"));
+      key[0]  = doc.getKey();
+    }
+    else
+    {
+      doc = col.insertAndGet(db.createDocumentFromString("id-1", createTblStr, "text/plain"));
+      key[0]  = doc.getKey();
+      assertEquals(createTblStr, new String(col.findOne(key[0]).getContentAsByteArray(), "UTF-8"));
+
+      doc = col.insertAndGet(db.createDocumentFromString("id-2", createTblStr, "text/plain"));
+      key[1]  = doc.getKey();
+      assertEquals(createTblStr, new String(col.findOne(key[1]).getContentAsByteArray(), "UTF-8"));
+    }
     
-    col.insert(db.createDocumentFromString("id-2", createTblStr, "text/plain"));
-    assertEquals(createTblStr, new String(col.findOne("id-2").getContentAsByteArray(), "UTF-8"));
-    
-    col.find().key("id-1").replaceOne(db.createDocumentFromString("{ \"sql\" : \"" + createTblStr + "\" }"));
-    doc = col.findOne("id-1");
-    assertEquals("{ \"sql\" : \"" + createTblStr + "\" }", new String(doc.getContentAsByteArray(), "UTF-8"));
+    if (isJDCSMode()) 
+    {
+      doc = col.find().key(key[0]).replaceOneAndGet(db.createDocumentFromString("{\"sql\" :\"" + createTblStr + "\"}"));
+      key[0]  = doc.getKey();
+    }
+    else
+    {
+      col.find().key(key[0]).replaceOne(db.createDocumentFromString("{\"sql\":\"" + createTblStr + "\"}"));
+    }
+    doc = col.findOne(key[0]);
+    assertEquals("{\"sql\":\"" + createTblStr + "\"}", new String(doc.getContentAsByteArray(), "UTF-8"));
     assertEquals("application/json", doc.getMediaType());
     
-    col.save(db.createDocumentFromString("id-1", createTblStr, "text/plain"));
-    doc = col.findOne("id-1");
-    assertEquals(createTblStr, new String(doc.getContentAsByteArray(), "UTF-8"));
-    assertEquals("text/plain", doc.getMediaType());
-    
-    col.save(db.createDocumentFromString("id-3", createTblStr, "text/plain"));
-    doc = col.findOne("id-3");
-    assertEquals(createTblStr, new String(doc.getContentAsByteArray(), "UTF-8"));
-    assertEquals("text/plain", doc.getMediaType());
-    
-    assertEquals(3, col.find().count());
+    if (!isJDCSMode()) 
+    {
+      col.save(db.createDocumentFromString("id-1", createTblStr, "text/plain"));
+      doc = col.findOne("id-1");
+      assertEquals(createTblStr, new String(doc.getContentAsByteArray(), "UTF-8"));
+      assertEquals("text/plain", doc.getMediaType());
+      
+      col.save(db.createDocumentFromString("id-3", createTblStr, "text/plain"));
+      doc = col.findOne("id-3");
+      assertEquals(createTblStr, new String(doc.getContentAsByteArray(), "UTF-8"));
+      assertEquals("text/plain", doc.getMediaType());
+      assertEquals(3, col.find().count());
+    }
+    else
+    {
+      assertEquals(1, col.find().count());
+    }
     col.find().remove();
 
     OracleDocument metaDoc2 = client.createMetadataBuilder()
       .keyColumnAssignmentMethod("CLIENT")
       .build();
-    OracleCollection col2 = dbAdmin.createCollection("testSQLInjection2", metaDoc2);
+    OracleCollection col2;
+    if (isJDCSMode())
+    {
+      col2 = dbAdmin.createCollection("testSQLInjection2", null);
+    } else
+    {
+      col2 = dbAdmin.createCollection("testSQLInjection2", metaDoc2);
+    }
 
     for (int i = 1; i <= 10; i++) {
-      col2.insertAndGet(db.createDocumentFromString("id-" + i, "{ \"dataValue\" : " + i + " }", null));
+      if (isJDCSMode()) 
+      {
+        doc = col2.insertAndGet(db.createDocumentFromString("{ \"dataValue\" : " + i + " }"));
+        key[i-1] = doc.getKey();
+      } else
+      {
+         doc = col2.insertAndGet(db.createDocumentFromString("id-" + i, "{ \"dataValue\" : " + i + " }", null));
+        key[i-1] = doc.getKey();
+      }     
     }
     
     String fStr = "{ \"$query\" : {\"dataValue\" : {\"$gt\" : \"" + createTblStr + "\" }}, \"$orderby\" : {\"d\" : -1} }";
@@ -128,20 +198,33 @@ public class test_OracleSodaSecurity extends SodaTestCase {
       }
     }
     else {
-      OracleCollection col3 = db.admin().createCollection(colName3.toUpperCase(), metaDoc3);
-      col3.insertAndGet(db.createDocumentFromString("id001", "{ \"dataValue\" : 1001 }", null));
-      assertEquals(1, col3.find().key("id001").count());
+      OracleCollection col3;
+      if (isJDCSMode()) 
+      {
+        col3 = db.admin().createCollection(colName3.toUpperCase(), null);
+        doc = col3.insertAndGet(db.createDocumentFromString("{ \"dataValue\" : 1001 }"));
+        key[2] = doc.getKey();
+      } else
+      {
+        col3 = db.admin().createCollection(colName3.toUpperCase(), metaDoc3);
+        doc = col3.insertAndGet(db.createDocumentFromString("id001", "{ \"dataValue\" : 1001 }", null));
+        key[2] = doc.getKey();
+      }
+      assertEquals(1, col3.find().key(key[2]).count());
       col3.admin().drop();
     }
     
     // Test with attacks via key column name
-    final String keyColumnStr = "CREATED_ON\" is not null) or 1=1 or (\"VERSION";
-    OracleDocument metaDoc4 = client.createMetadataBuilder().keyColumnSequenceName(keyColumnStr)
-      .keyColumnAssignmentMethod("CLIENT").build();
-    OracleCollection col4 = db.admin().createCollection("testSQLInjection4", metaDoc4);
-    col4.insertAndGet(db.createDocumentFromString("id001", "{ \"dataValue\" : 1001 }", null));
-    assertEquals(0, col4.find().key("idxxx").count());
-    col4.admin().drop();
+    if (!isJDCSMode()) 
+    {
+      final String keyColumnStr = "CREATED_ON\" is not null) or 1=1 or (\"VERSION";
+      OracleDocument metaDoc4 = client.createMetadataBuilder().keyColumnSequenceName(keyColumnStr)
+        .keyColumnAssignmentMethod("CLIENT").build();
+      OracleCollection col4 = db.admin().createCollection("testSQLInjection4", metaDoc4);
+      col4.insertAndGet(db.createDocumentFromString("id001", "{ \"dataValue\" : 1001 }", null));
+      assertEquals(0, col4.find().key("idxxx").count());
+      col4.admin().drop();
+    }
     
     // Test with attacks via version column name
     final String versionColumnStr = "ID\" is not null) or 1=1 or (\"CREATED_ON";
@@ -159,7 +242,14 @@ public class test_OracleSodaSecurity extends SodaTestCase {
       }
     }
     else {
-      col5 = db.admin().createCollection("testSQLInjection5", metaDoc5);
+      if (isJDCSMode())
+      {
+        col5 = db.admin().createCollection("testSQLInjection5", null);
+      } else
+      {
+        col5 = db.admin().createCollection("testSQLInjection5", metaDoc5);
+      }
+      
       doc = col5.insertAndGet(db.createDocumentFromString("{ \"dataValue\" : 1001 }"));
       assertEquals(0, col5.find().version("unknown version value").count());
       col5.admin().drop();
@@ -181,51 +271,79 @@ public class test_OracleSodaSecurity extends SodaTestCase {
       }
     }
     else {
-      col6 = db.admin().createCollection("testSQLInjection6", metaDoc6);
-      doc = col6.insertAndGet(db.createDocumentFromString("key1", "{ \"dataValue\" : 1001 }"));
-      assertEquals(0, ((OracleOperationBuilderImpl) col6.find().key("key1"))
+      if (isJDCSMode())
+      {
+        col6 = db.admin().createCollection("testSQLInjection6", null);
+        doc = col6.insertAndGet(db.createDocumentFromString("{ \"dataValue\" : 1001 }"));
+        key[0] = doc.getKey();
+      } else
+      {
+        col6 = db.admin().createCollection("testSQLInjection6", metaDoc6);
+        doc = col6.insertAndGet(db.createDocumentFromString("key1", "{ \"dataValue\" : 1001 }"));
+        key[0] = doc.getKey();
+      }     
+      
+      assertEquals(0, ((OracleOperationBuilderImpl) col6.find().key(key[0]))
         .lastModified("2017-05-17T07:29:56.825900Z").count());
       col6.admin().drop();
     }
     
     // Tests with attacks via version parameter
-    OracleDocument metaDoc7 = client.createMetadataBuilder()
-      .keyColumnAssignmentMethod("CLIENT")
-      .versionColumnMethod("NONE").tableName("SODATBL")
-      .build();
-    OracleCollection col7 = db.admin().createCollection("testSQLInjection7", metaDoc7);
-    
-    for (int i = 1; i <= 10; i++) {
-      col7.insertAndGet(db.createDocumentFromString("id-" + i, "{ \"value\" : " + i + " }", null));
+    if (!isJDCSMode())
+    {
+      OracleDocument metaDoc7 = client.createMetadataBuilder()
+        .keyColumnAssignmentMethod("CLIENT")
+        .versionColumnMethod("NONE").tableName("SODATBL")
+        .build();
+      OracleCollection col7 = db.admin().createCollection("testSQLInjection7", metaDoc7);
+      
+      for (int i = 1; i <= 10; i++) {
+        col7.insertAndGet(db.createDocumentFromString("id-" + i, "{ \"value\" : " + i + " }", null));
+      }
+      
+      assertEquals(0, col7.find().key("id-xxx").version("1 or 1=1 ").count());
+      assertEquals(0, col7.find().key("id-xxx").version("1') or (1=1) --").count());
+      
+      assertNull(col7.find().key("id-xxx").version("1 or 1=1 ").getOne());
+      assertNull(col7.find().key("id-xxx").version("1') or (1=1) --").getOne());
+      
+      assertEquals(0, col7.find().key("id-xxx").version("1') or 1=1").remove());
+      assertEquals(0, col7.find().key("id-xxx").version("1') or (1=1) --").remove());
+      
+      doc = db.createDocumentFromString("id-xxx", "{ \"value\" : \"replaced value\" }");
+      assertEquals(false, col7.find().key("id-xxx").version("1') or 1=1").replaceOne(doc));
+      assertEquals(false, col7.find().key("id-xxx").version("1') or (1=1) --").replaceOne(doc));
     }
-    
-    assertEquals(0, col7.find().key("id-xxx").version("1 or 1=1 ").count());
-    assertEquals(0, col7.find().key("id-xxx").version("1') or (1=1) --").count());
-    
-    assertNull(col7.find().key("id-xxx").version("1 or 1=1 ").getOne());
-    assertNull(col7.find().key("id-xxx").version("1') or (1=1) --").getOne());
-    
-    assertEquals(0, col7.find().key("id-xxx").version("1') or 1=1").remove());
-    assertEquals(0, col7.find().key("id-xxx").version("1') or (1=1) --").remove());
-    
-    doc = db.createDocumentFromString("id-xxx", "{ \"value\" : \"replaced value\" }");
-    assertEquals(false, col7.find().key("id-xxx").version("1') or 1=1").replaceOne(doc));
-    assertEquals(false, col7.find().key("id-xxx").version("1') or (1=1) --").replaceOne(doc));
-    
     
     // Test with sql injection attack via filter spec
     
     // Tests when sql injection attack happen in path filed
-    OracleDocument metaDoc8 = client.createMetadataBuilder().keyColumnAssignmentMethod("CLIENT").build();
-    OracleCollection col8 = db.admin().createCollection("testSQLInjection8", metaDoc8);
+    OracleDocument metaDoc8 = client.createMetadataBuilder().keyColumnAssignmentMethod("CLIENT").build();        
+    OracleCollection col8;
+    if (isJDCSMode())
+    {
+      col8 = db.admin().createCollection("testSQLInjection8", null);
+    } else
+    {
+      col8 = db.admin().createCollection("testSQLInjection8", metaDoc8);
+    }   
     String version2 = null;
-    
+
     for (int i = 1; i <= 10; i++) {
-      doc = col8.insertAndGet(db.createDocumentFromString("id-" + i, "{ \"d\" : " + i + " }"));
+      if (isJDCSMode())
+      {
+        doc = col8.insertAndGet(db.createDocumentFromString("{ \"d\" : " + i + " }"));
+        key[0] = doc.getKey();
+      } else
+      {
+        doc = col8.insertAndGet(db.createDocumentFromString("id-" + i, "{ \"d\" : " + i + " }"));
+        key[0] = doc.getKey();
+      }     
       if(i == 2)
-        version2 = doc.getVersion();
+        version2 = doc.getVersion();    
     }
-    
+
+
     OracleDocument filter = db.createDocumentFromString("{\"d==1||1==1||@.d\" : {\"$gt\" : 5}}");
     assertEquals(0, col8.find().filter(filter).count());
     
@@ -257,26 +375,36 @@ public class test_OracleSodaSecurity extends SodaTestCase {
     }
 
     // Tests when sql injection attack happen in "id" values
-    filter = db.createDocumentFromString("{\"$id\" : [\"id-0)or(1=1\"]}");
-    assertEquals(0, col8.find().filter(filter).count());
-    
-    filter = db.createDocumentFromString("{\"$id\" : [\"id-0')or(1=1\"]}");
-    assertEquals(0, col8.find().filter(filter).count());
-    
-    filter = db.createDocumentFromString("{\"$id\" : [\"id-0')or('1'='1\"]}");
-    assertEquals(0, col8.find().filter(filter).count());
-    
-    filter = db.createDocumentFromString("{\"$id\" : [\"id-1'))--\"]}");
-    assertEquals(0, col8.find().filter(filter).version(version2).count());
-    
+    if (!isJDCSMode())
+    {
+      filter = db.createDocumentFromString("{\"$id\" : [\"id-0)or(1=1\"]}");
+      assertEquals(0, col8.find().filter(filter).count());
+      
+      filter = db.createDocumentFromString("{\"$id\" : [\"id-0')or(1=1\"]}");
+      assertEquals(0, col8.find().filter(filter).count());
+      
+      filter = db.createDocumentFromString("{\"$id\" : [\"id-0')or('1'='1\"]}");
+      assertEquals(0, col8.find().filter(filter).count());
+      
+      filter = db.createDocumentFromString("{\"$id\" : [\"id-1'))--\"]}");
+      assertEquals(0, col8.find().filter(filter).version(version2).count());
+    }
     
     // Test with sql injection attack via spatial query
     OracleDocument metaDoc9 = client.createMetadataBuilder().keyColumnAssignmentMethod("CLIENT").build();
-    OracleCollection col9 = db.admin().createCollection("testSQLInjection9", metaDoc9);
-    
     String docStr1 = "{\"location\" : {\"type\": \"Point\", \"coordinates\": [33.7243,-118.1579]} }";
-    
-    doc = col9.insertAndGet(db.createDocumentFromString("id001", docStr1));
+    OracleCollection col9;
+    if (isJDCSMode())
+    {
+      col9 = db.admin().createCollection("testSQLInjection9", null);
+      doc = col9.insertAndGet(db.createDocumentFromString(docStr1));
+      key[0] = doc.getKey();
+    } else
+    {
+      col9 = db.admin().createCollection("testSQLInjection9", metaDoc9);
+      doc = col9.insertAndGet(db.createDocumentFromString("id001", docStr1));
+      key[0] = doc.getKey();
+    }  
     
     OracleDocument sptialSpec = db.createDocumentFromString(
         "{ \"location\" : { \"$near\" : {\n" +
